+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## In develop
 
 * Minor style fixes.
+* Fixed issue with change in group for user not being reflected in UI.
+* Removed unnecessary call to get all slides in channelController.
 
 ## 1.0.8